+++ .vscode/launch.json
@@ -11,7 +11,7 @@
             "preLaunchTask": "build",
             "program": "${workspaceFolder}/FPSOManagerApi_CS/bin/Debug/netcoreapp3.1/FPSOManagerApi_CS.dll",
             "args": [],
-            "cwd": "${workspaceFolder}",
+            "cwd": "${workspaceFolder}/FPSOManagerApi_CS",
             "stopAtEntry": false,
             "serverReadyAction": {
                 "action": "openExternally",
