+++ .vscode/tasks.json
@@ -7,7 +7,7 @@
             "type": "process",
             "args": [
                 "build",
-                "${workspaceFolder}/FPSOManagerApi_CS.csproj",
+                "${workspaceFolder}/FPSOManagerApi_CS/FPSOManagerApi_CS.csproj",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
@@ -19,7 +19,7 @@
             "type": "process",
             "args": [
                 "publish",
-                "${workspaceFolder}/FPSOManagerApi_CS.csproj",
+                "${workspaceFolder}/FPSOManagerApi_CS/FPSOManagerApi_CS.csproj",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
@@ -32,7 +32,7 @@
             "args": [
                 "watch",
                 "run",
-                "${workspaceFolder}/FPSOManagerApi_CS.csproj",
+                "${workspaceFolder}/FPSOManagerApi_CS/FPSOManagerApi_CS.csproj",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
