+++ ProjectTests/ProjectTests.csproj
@@ -1,20 +1,16 @@
 <Project Sdk="Microsoft.NET.Sdk">
-
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
-
     <IsPackable>false</IsPackable>
   </PropertyGroup>
-
   <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.5.0" />
-    <PackageReference Include="MSTest.TestAdapter" Version="2.1.0" />
-    <PackageReference Include="MSTest.TestFramework" Version="2.1.0" />
-    <PackageReference Include="coverlet.collector" Version="1.2.0" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.5.0"/>
+    <PackageReference Include="MSTest.TestAdapter" Version="2.1.0"/>
+    <PackageReference Include="MSTest.TestFramework" Version="2.1.0"/>
+    <PackageReference Include="coverlet.collector" Version="1.2.0"/>
+    <PackageReference Include="Moq" Version="4.14.6"/>
   </ItemGroup>
-
   <ItemGroup>
-    <ProjectReference Include="..\FPSOManagerApi_CS\FPSOManagerApi_CS.csproj" />
+    <ProjectReference Include="..\FPSOManagerApi_CS\FPSOManagerApi_CS.csproj"/>
   </ItemGroup>
-
 </Project>
